🎨 moved queries from mixins in other folder
