+++ src/application/queries/__init__.py
@@ -1,0 +1,15 @@
+from .auth import (
+    AuthQuery,
+)
+from .role import (
+    RoleQuery,
+)
+from .user import (
+    UserQuery,
+)
+
+__all__ = (
+    "UserQuery",
+    "AuthQuery",
+    "RoleQuery",
+)
