+++ src/application/queries/role.py
@@ -3,6 +3,7 @@ from typing import (
 )
 
 from sqlalchemy import (
+    Delete,
     Select,
     Update,
     select,
@@ -12,28 +13,32 @@ from sqlalchemy import (
 from src.application import (
     dto,
 )
-from src.domain.user import (
-    entity,
+from src.application.mixins.base import (
+    CRUDMixin,
+    UpdateSchemaT,
 )
 from src.infrastructure.database import (
     models,
 )
 
-from .base import (
-    CRUDMixin,
-)
 
-
-class UserQueryMixin(CRUDMixin[models.User, entity.User, dto.UserUpdate]):
+class RoleQuery(CRUDMixin[models.Role, dto.RoleCreate, dto.RoleUpdate]):
     def _get_query(self, *args: Any, **kwargs: Any) -> Select[tuple[Any]]:
         return select(self.model).filter(*args).filter_by(**kwargs)
 
     def _get_all_query(
             self, skip: int = 0, limit: int = 10, *args: Any, **kwargs: Any
     ) -> Select[tuple[Any]]:
-        return select(self.model).offset(skip).limit(limit).filter(*args).filter_by(**kwargs)
+        return (
+            select(self.model)
+            .offset(skip)
+            .limit(limit)
+            .filter(*args)
+            .filter_by(**kwargs)
+            .order_by(self.model.id)
+        )
 
-    def _update_query(self, pk: int, data_in: dto.UserUpdate, **kwargs: Any) -> Update:
+    def _update_query(self, pk: int, data_in: UpdateSchemaT, **kwargs: Any) -> Update:
         return (
             update(self.model)
             .where(self.model.id == pk)
@@ -42,9 +47,5 @@ class UserQueryMixin(CRUDMixin[models.User, entity.User, dto.UserUpdate]):
             .returning(self.model)
         )
 
-    def _delete_query(self, *args: Any, **kwargs: Any) -> Update:
-        return (
-            update(self.model)
-            .where(self.model.id == kwargs.get("id"))
-            .values(is_active=kwargs.get("is_active"))
-        )
+    def _delete_query(self, *args: Any, **kwargs: Any) -> Delete:
+        return Delete(self.model).filter(*args)
